🤖 refactor: make explorer a mux.explorer extension
- Register Explorer tab via mux.explorer builtin extension\n- Move ExplorerTab + fileExplorer helpers into src/extensions/explorer\n- Remove builtin:rightSidebarTabs (now redundant)\n\n---\n\n_Generated with `mux` • Model: openai:gpt-5.2 • Thinking: xhigh • Cost: \n\n<!-- mux-attribution: model=openai:gpt-5.2 thinking=xhigh costs=unknown -->
